Add base class for general logic of PrefabBrushes
