+++ Editor/Brushes/PrefabBrushes/BasePrefabBrush.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 71e49d8cb93a4e46b07eea7c8dd2186b
+timeCreated: 1584030887
