+++ Editor/Brushes/PrefabBrushes/PrefabBrush/PrefabBrush.cs
@@ -1,5 +1,4 @@
 using UnityEngine;
-using Object = UnityEngine.Object;
 
 namespace UnityEditor.Tilemaps
 {
@@ -8,7 +7,7 @@ namespace UnityEditor.Tilemaps
     /// </summary>
     [CreateAssetMenu(fileName = "Prefab brush", menuName = "Brushes/Prefab brush")]
     [CustomGridBrush(false, true, false, "Prefab Brush")]
-    public class PrefabBrush : GridBrush
+    public class PrefabBrush : BasePrefabBrush
     {
         /// <summary>
         /// The selection of Prefab to paint from
@@ -19,35 +18,22 @@ namespace UnityEditor.Tilemaps
         /// </summary>
         public bool m_ForceDelete;
         /// <summary>
-        /// Factor for distribution of choice of Prefabs to paint
-        /// </summary>
-        public float m_PerlinScale = 0.5f;
-        /// <summary>
-        /// Anchor Point of the Instantiated Prefab in the cell when painting
-        /// </summary>
-        public Vector3 m_Anchor = new Vector3(0.5f, 0.5f, 0.5f);
-        /// <summary>
         /// Paints Prefabs into a given position within the selected layers.
         /// The PrefabBrush overrides this to provide Prefab painting functionality.
         /// </summary>
         /// <param name="grid">Grid used for layout.</param>
         /// <param name="brushTarget">Target of the paint operation. By default the currently selected GameObject.</param>
         /// <param name="position">The coordinates of the cell to paint data to.</param>
+        
         public override void Paint(GridLayout grid, GameObject brushTarget, Vector3Int position)
         {
-            // Do not allow editing palettes
-            if (brushTarget.layer == 31 || brushTarget == null)
-                return;
-
-            GameObject prefab = m_Prefab;
+            
+           
+            Prefab = m_Prefab;
             var tileObject = GetObjectInCell(grid, brushTarget.transform, position);
             if (tileObject == null || tileObject.name != m_Prefab.name)
             {
-                GameObject instance = (GameObject)PrefabUtility.InstantiatePrefab(prefab);
-                Undo.MoveGameObjectToScene(instance, brushTarget.scene, "Paint Prefabs");
-                Undo.RegisterCreatedObjectUndo((Object)instance, "Paint Prefabs");
-                instance.transform.SetParent(brushTarget.transform);
-                instance.transform.position = grid.LocalToWorld(grid.CellToLocalInterpolated(position + m_Anchor));
+                base.Paint(grid, brushTarget, position);
             }
         }
 
@@ -61,35 +47,15 @@ namespace UnityEditor.Tilemaps
         /// <param name="position">The coordinates of the cell to erase data from.</param>
         public override void Erase(GridLayout grid, GameObject brushTarget, Vector3Int position)
         {
-            // Do not allow editing palettes
-            if (brushTarget.layer == 31)
-                return;
-
-            Transform erased = GetObjectInCell(grid, brushTarget.transform, position);
+            var erased = GetObjectInCell(grid, brushTarget.transform, position);
             if (erased == null)
             {
                 return;
             }
-            if (m_ForceDelete)
+            if (m_ForceDelete || (!m_ForceDelete && erased.gameObject.name == m_Prefab.name))
             {
-                Undo.DestroyObjectImmediate(erased.gameObject);
+                base.Erase(grid, brushTarget, position);
             }
-            else if (erased.gameObject.name == m_Prefab.name)
-            {
-                Undo.DestroyObjectImmediate(erased.gameObject);
-            }
-        }
-
-        private static Transform GetObjectInCell(GridLayout grid, Transform parent, Vector3Int position)
-        {
-            int childCount = parent.childCount;
-            for (int i = 0; i < childCount; i++)
-            {
-                Transform child = parent.GetChild(i);
-                if (position == grid.WorldToCell(child.position))
-                    return child;
-            }
-            return null;
         }
     }
 
@@ -97,19 +63,15 @@ namespace UnityEditor.Tilemaps
     /// The Brush Editor for a Prefab Brush.
     /// </summary>
     [CustomEditor(typeof(PrefabBrush))]
-    public class PrefabBrushEditor : GridBrushEditor
+    public class PrefabBrushEditor : BasePrefabBrushEditor
     {
         private SerializedProperty m_Prefab;
-        private SerializedProperty m_Anchor;
         private SerializedProperty m_ForceDelete;
-        private SerializedObject m_SerializedObject;
 
         protected override void OnEnable()
         {
             base.OnEnable();
-            m_SerializedObject = new SerializedObject(target);
             m_Prefab = m_SerializedObject.FindProperty("m_Prefab");
-            m_Anchor = m_SerializedObject.FindProperty("m_Anchor");
             m_ForceDelete = m_SerializedObject.FindProperty("m_ForceDelete");
         }
 
@@ -119,10 +81,10 @@ namespace UnityEditor.Tilemaps
         /// </summary>
         public override void OnPaintInspectorGUI()
         {
+            base.OnPaintInspectorGUI();
             m_SerializedObject.UpdateIfRequiredOrScript();
             EditorGUILayout.PropertyField(m_Prefab, true);
             EditorGUILayout.PropertyField(m_ForceDelete, true);
-            EditorGUILayout.PropertyField(m_Anchor);
             m_SerializedObject.ApplyModifiedPropertiesWithoutUndo();
         }
     }
