+++ Editor/Brushes/PrefabBrushes/PrefabRandomBrush/PrefabRandomBrush.cs
@@ -1,0 +1,70 @@
+using UnityEngine;
+
+namespace UnityEditor.Tilemaps
+{
+    /// <summary>
+    /// This Brush instances and places a randomly selected Prefabs onto the targeted location and parents the instanced object to the paint target. Use this as an example to quickly place an assorted type of GameObjects onto structured locations.
+    /// </summary>
+    [CreateAssetMenu(fileName = "Prefab Random brush", menuName = "Brushes/Prefab Random brush")]
+    [CustomGridBrush(false, true, false, "Prefab Random Brush")]
+    public class PrefabRandomBrush : BasePrefabBrush
+    {
+        private const float k_PerlinOffset = 100000f;
+        /// <summary>
+        /// The selection of Prefabs to paint from
+        /// </summary>
+        public GameObject[] m_Prefabs;
+        /// <summary>
+        /// Factor for distribution of choice of Prefabs to paint
+        /// </summary>
+        public float m_PerlinScale = 0.5f;
+        /// <summary>
+        /// Paints Prefabs into a given position within the selected layers.
+        /// The PrefabBrush overrides this to provide Prefab painting functionality.
+        /// </summary>
+        /// <param name="grid">Grid used for layout.</param>
+        /// <param name="brushTarget">Target of the paint operation. By default the currently selected GameObject.</param>
+        /// <param name="position">The coordinates of the cell to paint data to.</param>
+        public override void Paint(GridLayout grid, GameObject brushTarget, Vector3Int position)
+        {
+            var index = Mathf.Clamp(Mathf.FloorToInt(GetPerlinValue(position, m_PerlinScale, k_PerlinOffset) * m_Prefabs.Length), 0, m_Prefabs.Length - 1);
+            Prefab = m_Prefabs[index];
+            base.Paint(grid, brushTarget, position);
+        }
+        
+        private static float GetPerlinValue(Vector3Int position, float scale, float offset)
+        {
+            return Mathf.PerlinNoise((position.x + offset)*scale, (position.y + offset)*scale);
+        }
+    }
+
+    /// <summary>
+    /// The Brush Editor for a Prefab Brush.
+    /// </summary>
+    [CustomEditor(typeof(PrefabRandomBrush))]
+    public class PrefabRandomBrushEditor : BasePrefabBrushEditor
+    {
+        private PrefabRandomBrush prefabBrush { get { return target as PrefabRandomBrush; } }
+
+        private SerializedProperty m_Prefabs;
+        
+        protected override void OnEnable()
+        {
+            base.OnEnable();
+            m_Prefabs = m_SerializedObject.FindProperty("m_Prefabs");
+        }
+
+        /// <summary>
+        /// Callback for painting the inspector GUI for the PrefabBrush in the Tile Palette.
+        /// The PrefabBrush Editor overrides this to have a custom inspector for this Brush.
+        /// </summary>
+        public override void OnPaintInspectorGUI()
+        {
+            base.OnPaintInspectorGUI();
+            m_SerializedObject.UpdateIfRequiredOrScript();
+            prefabBrush.m_PerlinScale = EditorGUILayout.Slider("Perlin Scale", prefabBrush.m_PerlinScale, 0.001f, 0.999f);
+            EditorGUILayout.PropertyField(m_Prefabs, true);
+            m_SerializedObject.ApplyModifiedPropertiesWithoutUndo();
+        }
+    }
+}
